refactor : 필요없는 메소드 제거
