+++ README.md
@@ -106,14 +106,9 @@
 ### 예외
 - [x] 주말이나 공휴일에 출석할 시 예외를 발생시킨다
 
-## 크루 출석 객체 + 출석 상태 객체
+## 출석 객체 + 출석 상태 객체
 ### 기능
 - [x] 크루의 누적 출석, 지각 그리고 결석 횟수를 계산한다
-
-## 크루 출석부 + 경고 관리 객체
-### 기능
-- [ ] 크루의 제적 위험자 조회 결과를 조회할 수 있다
-- [x] 지각 3회당 결석 1번이다
 
 ## 시스템 요구사항
 ### 기능
