+++ src/main/java/AttendanceStatusCalculator.java
@@ -1,12 +1,11 @@
 import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 public class AttendanceStatusCalculator {
 
-    public Map<AttendanceStatus, Integer> calculateAllCount(List<Attendance> attendances) {
-        return attendances.stream()
+    public AttendanceStatusCounter calculateAllCount(List<Attendance> attendances) {
+        EnumMap<AttendanceStatus, Integer> attendanceStatusesCount = attendances.stream()
                 .map(Attendance::getAttendanceStatus)
                 .collect(Collectors.toMap(
                         status -> status,
@@ -14,5 +13,7 @@ public class AttendanceStatusCalculator {
                         Integer::sum,
                         () -> new EnumMap<>(AttendanceStatus.class)
                 ));
+
+        return AttendanceStatusCounter.create(attendanceStatusesCount);
     }
 }
